requirements.txt
# opencv-python
# Pillow
# pandas
# nbconvert
# jupyter
# nbformat
# matplotlib
# numpy
# tensorboard

absl-py==0.11.0
argon2-cffi==21.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==2.1.4
cycler==0.11.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.6
importlib-metadata==3.4.0
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
Pygments==2.10.0
pygobject==3.26.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.25
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
SecretStorage==2.3.1
Send2Trash==1.8.0
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.4.0
